requirements.txt
easydict==1.13
numpy==1.24.3
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==10.4.0
PyYAML==6.0.1
scikit_learn==1.5.2
tqdm==4.66.4
